requirements.txt
discord.py==2.1.0
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
psycopg2-binary
requests
sigfig
pygithub
python-dateutil
pytz
syllables
nltk
inflect
discord-pretty-help
python-dotenv